Packaging the python scripts as a python package
 The package opengrok_tools can be installed
 from the tarball, downloading all neccessary
 dependencies.
