+++ opengrok-tools/.gitignore
@@ -1,0 +1,4 @@
+env
+build
+dist
+opengrok_tools.egg-info
