+++ opengrok-tools/MANIFEST.in
@@ -1,0 +1,1 @@
+include src/main/python/README.md
