+++ opengrok-tools/src/main/python/opengrok_tools/all/__init__.py
@@ -1,0 +1,2 @@
+from .utils import *
+from .scm import *
