+++ opengrok-tools/src/main/python/opengrok_tools/all/scm/__init__.py
@@ -1,0 +1,6 @@
+from .cvs import CVSRepository
+from .git import GitRepository
+from .mercurial import MercurialRepository
+from .svn import SubversionRepository
+from .teamware import TeamwareRepository
+from .repository import Repository
