+++ opengrok-tools/src/main/python/opengrok_tools/all/utils/__init__.py
@@ -1,0 +1,21 @@
+from . import commands
+from . import command
+from . import opengrok
+from . import utils
+from . import filelock
+from . import hook
+from . import repofactory
+from . import webutil
+from . import readconfig
+
+__all__ = [
+    'opengrok',
+    'utils',
+    'command',
+    'commands',
+    'filelock',
+    'hook',
+    'repofactory',
+    'webutil',
+    'readconfig'
+]
