+++ opengrok-tools/src/main/python/opengrok_tools/all/utils/command.py
