+++ opengrok-tools/src/main/python/opengrok_tools/all/utils/readconfig.py
