+++ opengrok-tools/src/main/python/opengrok_tools/all/utils/utils.py
