+++ opengrok-tools/src/main/python/opengrok_tools/java.py
@@ -18,25 +18,24 @@
 # CDDL HEADER END
 
 #
-# Copyright (c) 2016, 2018, Oracle and/or its affiliates. All rights reserved.
+# Copyright (c) 2008, 2018, Oracle and/or its affiliates. All rights reserved.
+# Portions Copyright (c) 2017-2018, Chris Fraire <cfraire@me.com>.
 #
 
 import os
-import sys
 import argparse
-from java import Java, get_javaparser
+import sys
 import logging
 
+from all.utils.java import Java, get_javaparser
 
-"""
- Script for manipulating project groups
-"""
 
-if __name__ == '__main__':
+def main():
 
-    parser = argparse.ArgumentParser(description='Java wrapper for project '
-                                     'group manipulation',
+    parser = argparse.ArgumentParser(description='java wrapper',
                                      parents=[get_javaparser()])
+    parser.add_argument('-m', '--mainclass', required=True,
+                        help='Main class')
 
     args = parser.parse_args()
 
@@ -47,15 +46,16 @@ if __name__ == '__main__':
 
     logger = logging.getLogger(os.path.basename(sys.argv[0]))
 
-    cmd = Java(args.options, classpath=args.jar, java=args.java,
-               java_opts=args.java_opts,
-               main_class='org.opengrok.indexer.configuration.Groups',
-               logger=logger)
-    cmd.execute()
-    ret = cmd.getretcode()
+    java = Java(args.options, logger=logger, java=args.java,
+                jar=args.jar, java_opts=args.java_opts,
+                classpath=args.classpath, main_class=args.mainclass,
+                env_vars=args.environment)
+    java.execute()
+    ret = java.getretcode()
     if ret is None or ret != 0:
-        logger.error(cmd.getoutputstr())
-        logger.error("command failed (return code {})".format(ret))
+        logger.error(java.getoutputstr())
+        logger.error("java command failed (return code {})".format(ret))
         sys.exit(1)
-    else:
-        print(cmd.getoutputstr())
+
+if __name__ == '__main__':
+    main()
