+++ opengrok-tools/src/main/python/requirements.txt
@@ -1,0 +1,13 @@
+certifi==2018.8.24
+chardet==3.0.4
+idna==2.7
+JsonForm==0.0.2
+jsonschema==2.6.0
+JsonSir==0.0.2
+opengrok-tools==0.0.1
+Python-EasyConfig==0.1.7
+PyYAML==3.13
+requests==2.19.1
+Resource==0.2.1
+six==1.11.0
+urllib3==1.23
